router - added router, mapped history to redux store (#6)
