+++ src/containers/App/App.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables/colors.scss';
-@import '../../styles/variables/typography.scss';
+@import '../../styles/variables/colors';
+@import '../../styles/variables/typography';
 
 body {
   font-family: $font-family-main;
