+++ src/reducers/index.jsx
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
+import { connectRouter } from 'connected-react-router';
 import { loginReducer } from './loginReducer';
 
-export const rootReducer = combineReducers({
+export const mapToRootReducer = (history) => combineReducers({
   login: loginReducer,
+  router: connectRouter(history),
 });
